requirements.txt
asgiref==3.6.0
autopep8==2.0.1
boto3==1.26.107
botocore==1.29.107
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.10.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
whitenoise==6.4.0
